一些注解

+++ java8/c19/Observer/Watcher.java
@@ -9,6 +9,6 @@ import java.util.Observer;
 public class Watcher implements Observer {
     @Override
     public void update(Observable o, Object arg) {
-        System.out.println("update() called,count is " + ((Integer) arg).intValue());
+        System.out.println("update() called,count is " + arg);
     }
 }
